dev_requirements.txt
pytest
pytest-cov
pytest-asyncio
black
pylint
mypy
sphinx
sphinx_rtd_theme
myst_parser
azure-appconfiguration-provider
azure-monitor-opentelemetry
azure-monitor-events-extension
